fix linter

+++ usecase/repo_tx_registration.go
@@ -93,7 +93,7 @@ func (suite *UseCaseSuite) RegisterWorkerIdempotently(ctx context.Context, confi
 	}
 
 	// Give time for the tx to be included in a block
-	delay := int64(walletConfig.BlockDurationEstimated) * int64(walletConfig.RegistrationWaitingBlocks)
+	delay := int64(walletConfig.BlockDurationEstimated * float64(walletConfig.RegistrationWaitingBlocks))
 	log.Debug().Int64("delay", delay).Msg("Waiting to check registration status to be included in a block...")
 	if lib.DoneOrWait(ctx, delay) {
 		log.Error().Err(ctx.Err()).Str("rpc", queryNode.ServerAddress).Msg("Waiting to check registration status failed")
@@ -178,7 +178,7 @@ func (suite *UseCaseSuite) RegisterAndStakeReputerIdempotently(ctx context.Conte
 		}
 
 		// Give time for the tx to be included in a block
-		delay := int64(walletConfig.BlockDurationEstimated) * int64(walletConfig.RegistrationWaitingBlocks)
+		delay := int64(walletConfig.BlockDurationEstimated * float64(walletConfig.RegistrationWaitingBlocks))
 		log.Debug().Int64("delay", delay).Msg("Waiting to check registration status to be included in a block...")
 		if lib.DoneOrWait(ctx, delay) {
 			log.Error().Err(ctx.Err()).Str("rpc", queryNode.ServerAddress).Msg("Waiting to check registration status failed")
@@ -239,7 +239,7 @@ func (suite *UseCaseSuite) RegisterAndStakeReputerIdempotently(ctx context.Conte
 	}
 
 	// Give time for the tx to be included in a block
-	delay := int64(walletConfig.BlockDurationEstimated) * int64(walletConfig.RegistrationWaitingBlocks)
+	delay := int64(walletConfig.BlockDurationEstimated * float64(walletConfig.RegistrationWaitingBlocks))
 	log.Debug().Int64("delay", delay).Msg("Waiting to check stake status to be included in a block...")
 	if lib.DoneOrWait(ctx, delay) {
 		log.Error().Err(ctx.Err()).Str("rpc", queryNode.ServerAddress).Msg("Waiting to check stake status failed")
